+++ zuihou-backend/pom.xml
@@ -22,6 +22,7 @@
         <module>zuihou-jobs</module>
         <module>zuihou-gateway</module>
         <module>zuihou-api</module>
+        <module>zuihou-common</module>
         <!-- 以下两个模块为测试模块 -->
         <module>zuihou-demo</module>
         <module>zuihou-order</module>
@@ -119,6 +120,12 @@
                 <groupId>com.github.zuihou</groupId>
                 <artifactId>zuihou-zipkin-client-starter</artifactId>
                 <version>${commons.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.github.zuihou</groupId>
+                <artifactId>zuihou-common</artifactId>
+                <version>${project.version}</version>
             </dependency>
 
             <!--            <dependency>-->
