+++ zuihou-backend/zuihou-authority/zuihou-authority-controller/src/main/java/com/github/zuihou/authority/controller/auth/Test123Controller.java
@@ -1,0 +1,23 @@
+package com.github.zuihou.authority.controller.auth;
+
+
+import com.github.zuihou.authority.dto.core.OrgSaveDTO;
+import com.github.zuihou.authority.dto.core.OrgUpdateDTO;
+import com.github.zuihou.authority.entity.core.Org;
+import com.github.zuihou.authority.service.core.OrgService;
+import com.github.zuihou.base.BaseController2;
+import io.swagger.annotations.Api;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@Validated
+@RestController
+@RequestMapping("/test123")
+@Api(value = "test", tags = "test")
+public class Test123Controller extends BaseController2<OrgService, Long, Org, Org, OrgSaveDTO, OrgUpdateDTO> {
+
+
+}
