+++ zuihou-backend/zuihou-authority/zuihou-authority-server/src/main/java/com/github/zuihou/authority/config/mq/AuthorityMqAutoConfiguration.java
@@ -3,8 +3,8 @@ package com.github.zuihou.authority.config.mq;
 import com.alibaba.fastjson.JSONObject;
 import com.github.zuihou.authority.dto.auth.SystemApiScanSaveDTO;
 import com.github.zuihou.authority.service.auth.SystemApiService;
-import com.github.zuihou.common.constant.QueueConstants;
 import com.github.zuihou.context.BaseContextHandler;
+import com.github.zuihou.mq.constant.QueueConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
