+++ zuihou-backend/zuihou-common/pom.xml
@@ -3,36 +3,23 @@
          xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>zuihou-msgs</artifactId>
+        <artifactId>zuihou-backend</artifactId>
         <groupId>com.github.zuihou</groupId>
         <version>1.0-SNAPSHOT</version>
         <relativePath>../</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>zuihou-msgs-entity</artifactId>
+    <groupId>com.github.zuihou</groupId>
+    <artifactId>zuihou-common</artifactId>
     <name>${project.artifactId}</name>
-    <description>消息服务实体模块</description>
+    <description>公共模块</description>
 
     <dependencies>
-
         <dependency>
             <groupId>com.github.zuihou</groupId>
-            <artifactId>zuihou-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.github.zuihou</groupId>
-            <artifactId>zuihou-boot</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>io.springfox</groupId>
-            <artifactId>springfox-core</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.baomidou</groupId>
-            <artifactId>mybatis-plus</artifactId>
+            <artifactId>zuihou-core</artifactId>
         </dependency>
     </dependencies>
+
 </project>
