+++ zuihou-backend/zuihou-common/src/main/java/com/github/zuihou/common/constant/AttachmentType.java
