+++ zuihou-backend/zuihou-common/src/main/resources/j2cache/caffeine.properties
@@ -1,0 +1,25 @@
+#########################################
+# Caffeine configuration
+# \u6682\u65F6\u6CA1\u7528
+# [name] = size, xxxx[s|m|h|d]
+#########################################
+default=2000, 2h 
+captcha=1000, 5m
+menu=2000, 12h 
+user_menu=3000, 12h 
+resource=2000, 12h 
+user_resource=3000, 12h 
+role=2000, 12h 
+role_menu=4000, 12h 
+role_resource=4000, 12h 
+role_org=4000, 12h 
+user_role=4000, 12h
+area=5000, 48h
+area_all=50, 48h
+login_log_total=50, 2h 
+login_log_today=50, 2h 
+login_log_todayip=50, 2h 
+login_log_tenday=200, 2h 
+login_log_browser=50, 2h 
+login_log_system=50, 2h 
+
