+++ zuihou-backend/zuihou-demo/zuihou-demo-entity/pom.xml
@@ -16,6 +16,10 @@
     <dependencies>
         <dependency>
             <groupId>com.github.zuihou</groupId>
+            <artifactId>zuihou-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.github.zuihou</groupId>
             <artifactId>zuihou-boot</artifactId>
         </dependency>
         <dependency>
