+++ zuihou-backend/zuihou-gateway/zuihou-zuul-base/pom.xml
@@ -18,6 +18,10 @@
     <dependencies>
         <dependency>
             <groupId>com.github.zuihou</groupId>
+            <artifactId>zuihou-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.github.zuihou</groupId>
             <artifactId>zuihou-boot</artifactId>
             <exclusions>
                 <exclusion>
