+++ zuihou-commons/zuihou-mq-starter/src/main/java/com/github/zuihou/mq/constant/QueueConstants.java
@@ -1,4 +1,4 @@
-package com.github.zuihou.common.constant;
+package com.github.zuihou.mq.constant;
 
 /**
  * 队列名常量
